refactor(blocks-construct): resolve the sonarQube issue
