+++ src/features/dashboard/component/dashboard-user-activity-graph/dashboard-user-activity-graph.tsx
@@ -11,7 +11,7 @@ import {
 } from 'components/ui/select';
 import { ChartContainer, ChartTooltip } from 'components/ui/chart';
 import { chartConfig, chartData, daysOfWeek } from '../../services/dashboard-service';
-import { Payload, ValueType, NameType } from 'recharts/types/component/DefaultTooltipContent';
+import { DashboardUserActivityGraphTooltip } from './dashboard-user-activity-graph-tooltip';
 
 /**
  * DashboardUserActivityGraph component displays a bar chart visualizing user activity trends.
@@ -25,27 +25,6 @@ import { Payload, ValueType, NameType } from 'recharts/types/component/DefaultTo
  *
  * @returns {JSX.Element} - The rendered JSX component showing user activity trends over time with a selectable time period.
  */
-
-interface TooltipContentProps {
-  payload: Payload<ValueType, NameType>[] | undefined;
-  label: string;
-}
-
-const TooltipContent = ({ payload, label }: TooltipContentProps) => {
-  const { t } = useTranslation();
-  const data = payload?.[0]?.value;
-
-  if (!data) return null;
-
-  return (
-    <div className="flex flex-col gap-1 bg-white p-2 shadow-md rounded-[4px]">
-      <p className="text-sm text-high-emphasis">{label}:</p>
-      <p className="text-sm font-semibold text-medium-emphasis">
-        {data.toLocaleString()} {t('ACTIONS')}
-      </p>
-    </div>
-  );
-};
 
 export const DashboardUserActivityGraph = () => {
   const { t } = useTranslation();
@@ -90,7 +69,9 @@ export const DashboardUserActivityGraph = () => {
               tickFormatter={(value) => value.toLocaleString()}
             />
             <ChartTooltip
-              content={({ payload, label }) => <TooltipContent payload={payload} label={label} />}
+              content={({ payload, label }) => (
+                <DashboardUserActivityGraphTooltip payload={payload} label={label} />
+              )}
             />
             <Bar dataKey="noOfActions" fill="var(--color-noOfActions)" radius={4} />
           </BarChart>
